+++ ReBuff/Widgets/WidgetIcon.cs
@@ -221,7 +221,7 @@ namespace ReBuff.Widgets
 
                     else
                     {
-                        ushort icon = style.IconOption switch
+                        uint icon = style.IconOption switch
                         {
                             0 => data.Icon,
                             1 => style.CustomIcon,
